backend/pyproject.toml
[project]
name = "course-platform-backend"
version = "0.1.0"
description = "Backend da plataforma de cursos — FastAPI + SQLAlchemy + Alembic"
requires-python = ">=3.12"

dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    # ORM + driver PostgreSQL assíncrono
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    # Migrações
    "alembic>=1.13.0",
    # Configuração via variáveis de ambiente
    "pydantic-settings>=2.4.0",
    # Upload multipart (vídeos)
    "python-multipart>=0.0.12",
    # MinIO / S3
    "boto3>=1.35.0",
    # HTTP client async (Drive API)
    "httpx>=0.27.0",
    # Extração de metadados do YouTube sem API key
    "yt-dlp>=2024.1.0",
    # Notion SDK (Passo 7)
    "notion-client>=2.2.1",
]

[project.optional-dependencies]
dev = [
    # Testes
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    # Driver SQLite assíncrono para testes unitários (sem Docker)
    "aiosqlite>=0.20.0",
    # Linting e formatação
    "ruff>=0.6.0",
    # Type checking
    "mypy>=1.11.0",
]

# ---------------------------------------------------------------
# Ruff: linting + formatação (substitui flake8 + black + isort)
# ---------------------------------------------------------------
[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["B008"]  # FastAPI usa Depends() em parâmetros de função

# ---------------------------------------------------------------
# Pytest
# ---------------------------------------------------------------
[tool.pytest.ini_options]
asyncio_mode = "auto"           # todos os testes async sem decorator extra
testpaths = ["tests"]
addopts = "--cov=app --cov-report=term-missing -v"

# ---------------------------------------------------------------
# MyPy
# ---------------------------------------------------------------
[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
